[Dennis Chan] Super smart parking boy test case

+++ src/main/java/com/oocl/ParkingLot.java
@@ -46,6 +46,6 @@ public class ParkingLot {
     }
 
     public double getPositionRate(){
-        return getRemainPosition()/capacity;
+        return (double)getRemainPosition()/capacity;
     }
 }
